refactor(shared): типы вынесены в отдельный файл
типы вынесены в отдельный файл
